Move StudentSection and MentorSection into their own modules
