+++ csm_web/frontend/src/components/Course.js
@@ -8,7 +8,7 @@ import ClockIcon from "../../static/frontend/img/clock.svg";
 import CheckCircle from "../../static/frontend/img/check_circle.svg";
 import XCircle from "../../static/frontend/img/x_circle.svg";
 import Modal, { ModalCloser } from "./Modal";
-import { SPACETIME_SHAPE } from "./Section";
+import { SPACETIME_SHAPE } from "../utils/types";
 
 const DAY_OF_WEEK_ABREVIATIONS = Object.freeze({
   Mon: "M",
