+++ csm_web/frontend/src/components/MentorSection.js
@@ -1,0 +1,4 @@
+import React from "react";
+export default function MentorSection() {
+  return <div>I am a mentor</div>;
+}
