+++ csm_web/frontend/src/utils/types.js
@@ -1,0 +1,5 @@
+import PropTypes from "prop-types";
+export const SPACETIME_SHAPE = PropTypes.shape({
+  location: PropTypes.string.isRequired,
+  time: PropTypes.string.isRequired
+});
